remove non-distributable automake cruft
until the source tree is set up so that `make dist` can generate a
proper distribution with local copies of the relevant autotools, users
won't be able to do the ./configure && make dance from what is
currently provided. Instead provide a bootstrap.sh script to take care
of creating the correct files.