hackish fix for ns_mysql_header hack
The "namespace ns_mysql_header" idea not a good plan. See comments in
the code, but in brief, system headers included for the first time via
mysql.h will be placed in this namespace and will be unavailable to the
global namespace. This patch just fixes a specific problem on a
specific OS, but it is very likely the problem will arise again with a
different OS or changes to the system headers included by mysql.h. It
would be a much better thing to figure out how to address the
underlying problem that was "solved" with this namespace, but I don't
quite know exactly what that was at this point.