Fix issues with implicit SQL default values
It is a SQL error to perform an INSERT that does not specify the value
of a "NOT NULL" column if the column has no default value. Certain
versions of mysql allow this and only generate warnings. Later versions
are more stringent.

In this code, there are lots of columns that seem to be de facto
optional but  are NOT NULL and have no default value. This is rather
odd table design, but assuming that the NOT NULLs are there for a
reason, then the solutions are to either provide proper defaults in the
schema (NB: text and blob cannot have default values!), or to make sure
that all INSERT queries provide all required column values. Another
possibility is to set the mysql session mode to be as permissive as
possible.

I have taken the route of changing the code, though the most elegant
would be to fix the table design to have proper defaults and/or allow
NULL column entries to signify "no entry" as is usually done.