Add discoverable XML submission to browser
Not sure why the only way to submit XML experiment job specifications
in the worm browser is drag-and-drop or clipboard actions... this is
very much not user-discoverable. (We only figured it out by grepping
for where the XML files get parsed.)

The undocumented nature of the specification for those XML files is of
course an issue for another time.