Store metadata in TIFFTAGS
Trying to store custom metadata in TIFF files in a way that persists
across file editing is an exercise in futility. XMP is only supported
by Photoshop, and the GIMP only supports the ImageDescription tag and
blanks the others. And delightfully, ImageJ overwrites the
ImageDescription tag.

The most-compatible of all of these bad options in the TIFFTAGS one.