Throttle FLTK idle function to not use 100% CPU
Defining an idle function makes it run as frequently as possible. This
was causing the worm browser to chew up 100% CPU (on a mac at least).
This seems to be less of a problem on windows? Not sure why. If the
idle functions need to run more often that 20 fps, that can be changed
of course. Or is there a reason they need to constantly loop?
Responsiveness of menus etc. is unaffected even at 20 fps.